requirements.txt
Django==5.1.3
djangorestframework==3.15.2
pymongo==4.10.1
django-cors-headers==4.6.0